+++ client/src/components/modals/admin/10-codes/Edit10CodeModal.tsx
@@ -1,97 +1,81 @@
 import * as React from "react";
 import { connect } from "react-redux";
-import { add10Code } from "../../../lib/actions/admin";
-import Select from "../../../components/select";
-import Code10 from "../../../interfaces/Code10";
-import useDocTitle from "../../../hooks/useDocTitle";
-import State from "../../../interfaces/State";
-import Modal from "..";
-import { ModalIds } from "../../../lib/types";
-import { modal } from "../../../lib/functions";
-
-export const options = [
-  {
-    value: "leo",
-    label: "LEO",
-  },
-  {
-    value: "dispatch",
-    label: "Dispatch",
-  },
-  {
-    value: "ems_fd",
-    label: "EMS-FD",
-  },
-];
-
-export const shouldDoOptions = [
-  {
-    value: "set_status",
-    label: window.lang.admin.set_status,
-  },
-  {
-    value: "set_off_duty",
-    label: window.lang.admin.set_off_duty,
-  },
-];
-
-export const colorOptions = [
-  {
-    value: "btn-danger",
-    label: "Red",
-  },
-  {
-    value: "btn-secondary",
-    label: "Normal",
-  },
-];
+import { update10Code } from "../../../../lib/actions/admin";
+import Select from "../../../../components/select";
+import State from "../../../../interfaces/State";
+import Code10 from "../../../../interfaces/Code10";
+import {
+  colorOptions,
+  shouldDoOptions,
+  options,
+} from "../../../../components/modals/admin/10-codes/Create10CodeModal";
+import useDocTitle from "../../../../hooks/useDocTitle";
+import { modal } from "../../../../lib/functions";
+import Modal from "../..";
+import { ModalIds } from "../../../../lib/types";
 
 interface Props {
-  codesLength: number;
-  add10Code: (data: Partial<Code10>) => Promise<boolean>;
+  update10Code: (id: string, data: Partial<Code10>) => Promise<boolean>;
+  code: Code10 | null;
+  codes: Code10[];
 }
 
-const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
+const Edit10CodeModal: React.FC<Props> = ({ update10Code, code: data, codes }) => {
   const [code, setCode] = React.useState("");
-  const [whatPages, setWhatPages] = React.useState([]);
+  const [whatPages, setWhatPages] = React.useState<Code10["what_pages"]>([]);
   const [color, setColor] = React.useState("");
   const [shouldDo, setShouldDo] = React.useState("");
   const [position, setPosition] = React.useState(0);
-  useDocTitle(window.lang.codes.add_10_code);
+  useDocTitle(window.lang.codes.edit_10_code);
+
+  const value = {
+    value: colorOptions.find((clr) => clr.value === color)?.value,
+    label: colorOptions.find((clr) => clr.value === color)?.label,
+  };
+  const shouldDoValue = {
+    value: shouldDoOptions.find((option) => option.value === shouldDo)?.value,
+    label: shouldDoOptions.find((option) => option.value === shouldDo)?.label,
+  };
+
+  React.useEffect(() => {
+    if (!data) return;
+
+    setCode(data.code);
+    setWhatPages(data.what_pages);
+    setColor(data.color);
+    setShouldDo(data.should_do);
+    setPosition(data.position);
+  }, [data]);
 
   const length = React.useCallback(() => {
     const arr = [];
 
-    for (let i = 0; i <= codesLength + 1; i++) {
+    for (let i = 0; i <= codes?.length + 1; i++) {
       arr.push(i);
     }
 
     return arr.map((v) => ({ value: v, label: `${window.lang.codes.position}: ${++v}` }));
-  }, [codesLength]);
+  }, [codes?.length]);
 
   async function onSubmit(e: React.FormEvent) {
     e.preventDefault();
+    if (!data?.id) return;
 
-    const added = await add10Code({
+    const updated = await update10Code(data?.id, {
       code,
-      color: color,
       what_pages: whatPages,
+      color: color,
       should_do: shouldDo,
       position,
     });
 
-    if (added === true) {
-      setCode("");
-      setWhatPages([]);
-      setColor("");
-      setShouldDo("");
-      setPosition(0);
-      modal(ModalIds.Create10Code).hide();
+    if (updated === true) {
+      modal(ModalIds.Edit10Code).hide();
     }
   }
 
   return (
-    <Modal size="lg" title={window.lang.codes.add_10_code} id={ModalIds.Create10Code}>
+    <Modal size="lg" id={ModalIds.Edit10Code} title={window.lang.codes.edit_10_code}>
       <form onSubmit={onSubmit}>
         <div className="modal-body">
           <div className="mb-3">
@@ -110,10 +94,12 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
               {window.lang.codes.code_where}
             </label>
             <Select
+              isClearable={false}
               closeMenuOnSelect={false}
               isMulti
               options={options}
               onChange={(v: any) => setWhatPages(v)}
+              value={whatPages}
             />
           </div>
           <div className="mb-3">
@@ -121,10 +107,12 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
               {window.lang.codes.select_color}
             </label>
             <Select
+              isClearable={false}
               closeMenuOnSelect
               isMulti={false}
               options={colorOptions}
-              onChange={(v) => setColor(v.value)}
+              onChange={(v: any) => setColor(v.value)}
+              value={value as any}
             />
           </div>
           <div className="mb-3">
@@ -132,10 +120,12 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
               {window.lang.codes.what_it_do}
             </label>
             <Select
+              isClearable={false}
               closeMenuOnSelect
               isMulti={false}
               options={shouldDoOptions}
-              onChange={(v) => setShouldDo(v.value)}
+              onChange={(v: any) => setShouldDo(v.value)}
+              value={shouldDoValue as any}
             />
           </div>
           <div className="mb-3">
@@ -143,6 +133,11 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
               {window.lang.codes.position}
             </label>
             <Select
+              isClearable={false}
+              value={{
+                label: `${window.lang.codes.position}: ${position}`,
+                value: `${position}`,
+              }}
               closeMenuOnSelect
               isMulti={false}
               options={length() as any}
@@ -156,7 +151,7 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
             {window.lang.global.close}
           </button>
           <button type="submit" className="btn btn-primary">
-            {window.lang.codes.add_code}
+            {window.lang.global.update}
           </button>
         </div>
       </form>
@@ -165,7 +160,7 @@ const Create10CodeModal: React.FC<Props> = ({ codesLength, add10Code }) => {
 };
 
 const mapToProps = (state: State) => ({
-  codesLength: state.admin?.codes?.length,
+  codes: state.admin.codes,
 });
 
-export default connect(mapToProps, { add10Code })(Create10CodeModal);
+export default connect(mapToProps, { update10Code })(Edit10CodeModal);
